+++ tests/unittests/test_baselines.py
@@ -5,7 +5,7 @@ Compare current results to baseline
 import sciris as sc
 import covasim as cv
 
-do_save = True
+do_save = False
 baseline_filename  = sc.thisdir(__file__, 'baseline.json')
 benchmark_filename = sc.thisdir(__file__, 'benchmark.json')
 baseline_key = 'summary'
@@ -17,7 +17,8 @@ def save_baseline(do_save=do_save):
 
     sim = cv.Sim(verbose=0)
     sim.run()
-    sim.to_json(filename=baseline_filename, keys=baseline_key)
+    if do_save:
+        sim.to_json(filename=baseline_filename, keys=baseline_key)
 
     print('Done.')
 
@@ -28,8 +29,7 @@ def test_baseline():
     ''' Compare the current default sim against the saved baseline '''
 
     # Load existing baseline
-    filepath = sc.makefilepath(filename=baseline_filename, folder=sc.thisdir(__file__))
-    baseline = sc.loadjson(filepath)
+    baseline = sc.loadjson(baseline_filename)
     old = baseline[baseline_key]
 
     # Calculate new baseline
@@ -82,7 +82,8 @@ def test_baseline():
 def test_benchmark(do_save=do_save):
     ''' Compare benchmark performance '''
 
-    print('Updating benchmark...')
+    print('Running benchmark...')
+    previous = sc.loadjson(benchmark_filename)
 
     # Create the sim
     sim = cv.Sim(verbose=0)
@@ -109,6 +110,12 @@ def test_benchmark(do_save=do_save):
                 'n_days':   sim['n_days'],
                 },
             }
+
+    print('Previous benchmark:')
+    sc.pp(previous)
+
+    print('\nNew benchmark:')
+    sc.pp(json)
 
     if do_save:
         sc.savejson(filename=benchmark_filename, obj=json, indent=2)