+++ tests/unittests/test_simulation_parameter.py
@@ -18,7 +18,6 @@ class SimulationParameterTests(CovaSimTest):
         super().tearDown()
         pass
 
-    @unittest.skip("Need to construct a population now")
     def test_population_size(self):
         """
         Set population size to vanilla (1234)
@@ -26,12 +25,19 @@ class SimulationParameterTests(CovaSimTest):
 
         Depends on run default simulation
         """
-        self.set_microsim()
         TPKeys = TestProperties.ParameterKeys.SimulationKeys
+        pop_2_one_day = {
+            TPKeys.population_scaling_factor: 1,
+            TPKeys.number_simulated_days: 1,
+            TPKeys.number_agents: 2,
+            TPKeys.number_contacts: {'a': 1},
+            TPKeys.initial_infected_count: 0
+        }
         pop_10_one_day = {
             TPKeys.population_scaling_factor: 1,
             TPKeys.number_simulated_days: 1,
             TPKeys.number_agents: 10,
+            TPKeys.number_contacts: {'a': 4},
             TPKeys.initial_infected_count: 0
         }
         pop_123_one_day = {
@@ -46,54 +52,101 @@ class SimulationParameterTests(CovaSimTest):
             TPKeys.number_agents: 1234,
             TPKeys.initial_infected_count: 0
         }
+        pop_10k_one_day = {
+            TPKeys.population_scaling_factor: 1,
+            TPKeys.number_simulated_days: 1,
+            TPKeys.number_agents: 1e4,
+            TPKeys.initial_infected_count: 0
+        }
+        pop_100k_one_day = {
+            TPKeys.population_scaling_factor: 1,
+            TPKeys.number_simulated_days: 1,
+            TPKeys.number_agents: 1e5,
+            TPKeys.initial_infected_count: 0
+        }
+        pop_200k_one_day = {
+            TPKeys.population_scaling_factor: 1,
+            TPKeys.number_simulated_days: 1,
+            TPKeys.number_agents: 2e5,
+            TPKeys.initial_infected_count: 0
+        }
+        pop_400k_one_day = {
+            TPKeys.population_scaling_factor: 1,
+            TPKeys.number_simulated_days: 1,
+            TPKeys.number_agents: 4e5,
+            TPKeys.initial_infected_count: 0
+        }
+        pop_1M_one_day = {
+            TPKeys.population_scaling_factor: 1,
+            TPKeys.number_simulated_days: 1,
+            TPKeys.number_agents: 1e6,
+            TPKeys.initial_infected_count: 0
+        }
+        self.run_sim(pop_2_one_day)
+        pop_2_pop = self.get_day_zero_channel_value()
         self.run_sim(pop_10_one_day)
         pop_10_pop = self.get_day_zero_channel_value()
         self.run_sim(pop_123_one_day)
         pop_123_pop = self.get_day_zero_channel_value()
         self.run_sim(pop_1234_one_day)
         pop_1234_pop = self.get_day_zero_channel_value()
+
+        self.run_sim(pop_10k_one_day)
+        pop_10k_pop = self.get_day_zero_channel_value()
+        self.run_sim(pop_100k_one_day)
+        pop_100k_pop = self.get_day_zero_channel_value()
+        # self.run_sim(pop_200k_one_day)
+        # pop_200k_pop = self.get_day_zero_channel_value()
+
+        # self.run_sim(pop_400k_one_day)
+        # pop_400k_pop = self.get_day_zero_channel_value()
+        # self.run_sim(pop_1M_one_day)
+        # pop_1M_pop = self.get_day_zero_channel_value()
+
+        self.assertEqual(pop_2_pop, pop_2_one_day[TPKeys.number_agents])
         self.assertEqual(pop_10_pop, pop_10_one_day[TPKeys.number_agents])
         self.assertEqual(pop_123_pop, pop_123_one_day[TPKeys.number_agents])
         self.assertEqual(pop_1234_pop, pop_1234_one_day[TPKeys.number_agents])
+
+        self.assertEqual(pop_10k_pop, pop_10k_one_day[TPKeys.number_agents])
+        self.assertEqual(pop_100k_pop, pop_100k_one_day[TPKeys.number_agents])
+        # self.assertEqual(pop_200k_pop, pop_200k_one_day[TPKeys.number_agents])
+
+        # self.assertEqual(pop_400k_pop, pop_400k_one_day[TPKeys.number_agents])
+        # self.assertEqual(pop_1M_pop, pop_1M_one_day[TPKeys.number_agents])
         pass
 
-    @unittest.skip("See GH 162")
     def test_population_size_ranges(self):
         """
         Intent is to test zero, negative, and excessively large pop sizes
         """
-        self.set_microsim()
-        pop_zero_one_day = {
+        pop_neg_one_day = {
             TPKeys.population_scaling_factor: 1,
             TPKeys.number_simulated_days: 1,
-            TPKeys.number_agents: 0,
+            TPKeys.number_agents: -10,
             TPKeys.initial_infected_count: 0
         }
         with self.assertRaises(ValueError) as context:
-            self.run_sim(pop_zero_one_day)
-            pass
+            self.run_sim(pop_neg_one_day)
         error_message = str(context.exception)
-        self.assertIn("n", error_message) # Not awesome but the parameter is 'n'
-        pass
+        self.assertIn("negative", error_message)
 
-    @unittest.skip("See GH 162")
-    def test_negative_infected_count(self):
-        """
-        Test negative infected count
-        """
-        self.set_smallpop_hightransmission()
-        negative_infected_count = {
+        pop_zero_one_day = {
             TPKeys.population_scaling_factor: 1,
-            TPKeys.initial_infected_count: -1
+            TPKeys.number_simulated_days: 100,
+            TPKeys.number_agents: 0,
+            TPKeys.initial_infected_count: 0
         }
-        with self.assertRaises(ValueError) as context:
-            self.run_sim(negative_infected_count)
-            pass
-        error_message = str(context.exception)
-        self.assertIn('pop_infected', error_message)
+        self.run_sim(pop_zero_one_day)
+        self.assertEqual(self.simulation_result['results'][ResKeys.susceptible_at_timestep][-1], 0)
+        self.assertEqual(self.simulation_result['results'][ResKeys.susceptible_at_timestep][0], 0)
+
+        # with self.assertRaises(Exception) as context:
+        #     self.run_sim(pop_zero_one_day)
+        # error_message = str(context.exception)
+        # self.assertIn("population", error_message)
         pass
 
-    @unittest.skip("Need to construct a population now")
     def test_population_scaling(self):
         """
         Scale population vanilla (x10) compare
@@ -101,7 +154,6 @@ class SimulationParameterTests(CovaSimTest):
 
         Depends on population_size
         """
-        self.set_microsim()
         scale_1_one_day = {
             TPKeys.population_scaling_factor: 1,
             TPKeys.number_simulated_days: 1